requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==1.4.49
asyncpg==0.27.0
alembic==1.11.2
passlib==1.7.4
python-jose==3.3.0
pydantic==1.10.12
google-generativeai==0.3.0

jose~=1.0.0
python-dotenv~=1.0.1